+++ OgreMain/src/OgreStringConverter.cpp
@@ -58,7 +58,7 @@ namespace Ogre {
         return stream.str();
     }
     //-----------------------------------------------------------------------
-    String StringConverter::toString(unsigned int val, 
+    String StringConverter::toString(size_t val, 
         unsigned short width, char fill, std::ios::fmtflags flags)
     {
         StringUtil::StrStreamType stream;
